Enhance and remove native base components from SearchableHeader
