+++ app/components/schools/SchoolNavigationHeader.js
@@ -1,12 +1,26 @@
 import React from 'react'
+import {StyleSheet} from 'react-native'
 
 import SegmentView from './segment_view'
-import CustomNavigationHeader from '../shared/CustomNavigationHeader';
+import SearchableHeader from '../shared/searchableHeaders/SearchableHeader';
 
 const SchoolNavigationHeader = (props) => {
-  return <CustomNavigationHeader title='គ្រឹះស្ថានសិក្សា'>
+  return <SearchableHeader title="គ្រឹះស្ថានសិក្សា" placeholder="ស្វែងរកគ្រឹះស្ថានសិក្សា" containerStyle={[styles.header, {paddingTop: 4}]}
+            searchedText={props.searchText}
+            setSearchedText={props.setSearchedText}
+         >
             <SegmentView activePage={props.activePage} setContent={(active) => props.setContent(active)}/>
-         </CustomNavigationHeader>
+         </SearchableHeader>
 }
+
+const styles = StyleSheet.create({
+  header: {
+    backgroundColor: 'white',
+    elevation: 1,
+    flexDirection: 'column',
+    height: 'auto',
+    paddingBottom: 16,
+  },
+})
 
 export default SchoolNavigationHeader
