+++ app/components/shared/searchableHeaders/SearchableHeader.js
@@ -1,34 +1,32 @@
 import React from 'react'
-import {StyleSheet, TouchableOpacity} from 'react-native'
+import {TouchableOpacity, View, TextInput} from 'react-native'
 import {Appbar} from 'react-native-paper'
-import { Input, Icon, Item } from 'native-base';
+import Icon from 'react-native-vector-icons/MaterialIcons';
 
 import {BackButton} from '../..'
 import {goBack} from '../../../hooks/RootNavigation'
 import { FontSetting } from '../../../assets/style_sheets/font_setting'
 import {getStyleOfOS, getStyleOfDevice} from '../../../utils/responsive_util'
-import {navHeaderPaddingTop} from '../../../constants/component_constant'
 import {FontFamily} from '../../../themes/font';
 import color from '../../../themes/color'
 
 const SearchableHeader = (props) => {
   const [searchVisible, setSearchVisible] = React.useState(false)
   const searchBox = () => {
-    return (
-      <Item style={{flex: 1, borderRadius: 12, borderBottomWidth: 0, marginLeft: 6, marginRight: 16, backgroundColor: color.paleGray, height: 40}}>
-        <Input autoCorrect={false} autoFocus={true} placeholder={props.placeholder}
-          value={props.searchedText}
-          onChangeText={(text) => props.setSearchedText(text)}
-          placeholderTextColor={color.grayColor}
-          style={{fontSize: FontSetting.nav_title, fontFamily: FontFamily.regular, paddingLeft: 12, marginTop: getStyleOfOS(-3, 2)}}
-        />
-        { !!props.searchedText &&
-          <TouchableOpacity style={styles.iconContainer} onPress={() => props.setSearchedText('')}>
-            <Icon active name='close' style={{color: color.lightBlackColor}} />
-          </TouchableOpacity>
-        }
-      </Item>
-    )
+    return <View style={{flex: 1, flexDirection: 'row', borderRadius: 12, marginLeft: 6, marginRight: 16, backgroundColor: color.paleGray, height: 40, marginVertical: 4}}>
+              <TextInput
+                autoCorrect={false} autoFocus={true} placeholder={props.placeholder}
+                value={props.searchedText}
+                onChangeText={(text) => props.setSearchedText(text)}
+                placeholderTextColor={color.grayColor}
+                style={{flex: 1, fontFamily: FontFamily.regular, fontSize: FontSetting.text, height: 40, paddingTop: 0, paddingBottom: 0, paddingLeft: 16}}
+              />
+              { !!props.searchedText &&
+                <TouchableOpacity onPress={() => props.setSearchedText('')} style={{width: 48, height: '100%', justifyContent: 'center', alignItems: 'center'}}>
+                  <Icon name='close' size={24} style={{color: color.lightBlackColor, width: 24, paddingRight: 0}} />
+                </TouchableOpacity>
+              }
+           </View>
   }
 
   const renderTitle = () => {
@@ -40,7 +38,7 @@ const SearchableHeader = (props) => {
 
   const handlePressBack = () => {
     if (searchVisible) {
-      props.setSearchedText('')
+      !!props.setSearchedText && props.setSearchedText('')
       setSearchVisible(false)
       return
     }
@@ -50,27 +48,13 @@ const SearchableHeader = (props) => {
 
   return (
     <Appbar.Header style={[{backgroundColor: 'white', elevation: 1}, props.containerStyle]}>
-      <BackButton onPress={() => handlePressBack()} />
-      {searchVisible ? searchBox() : renderTitle()}
+      <View style={{flexDirection: 'row', alignItems: 'center'}}>
+        <BackButton onPress={() => handlePressBack()} />
+        {searchVisible ? searchBox() : renderTitle()}
+      </View>
+      {props.children}
     </Appbar.Header>
   )
 }
-
-const styles = StyleSheet.create({
-  header: {
-    backgroundColor: 'white',
-    elevation: 1,
-    flexDirection: 'column',
-    height: 'auto',
-    paddingBottom: 16,
-    paddingTop: navHeaderPaddingTop
-  },
-  title: {
-    color: 'black',
-    fontSize: FontSetting.nav_title,
-    marginTop: getStyleOfOS(0, -4),
-    paddingHorizontal: 16,
-  }
-})
 
 export default SearchableHeader
